requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
requests==2.32.3
pydantic==2.10.6
python-dotenv==1.0.1
pytest==8.3.4
